Organizing offchain utilities
