+++ dkg-runtime-primitives/src/lib.rs
@@ -6,6 +6,7 @@ pub mod mmr;
 pub mod proposal;
 pub mod traits;
 pub mod utils;
+pub mod offchain;
 
 use crypto::AuthorityId;
 pub use ethereum::*;
@@ -50,22 +51,7 @@ pub const DKG_ENGINE_ID: sp_runtime::ConsensusEngineId = *b"WDKG";
 // Key type for DKG keys
 pub const KEY_TYPE: sp_application_crypto::KeyTypeId = sp_application_crypto::KeyTypeId(*b"wdkg");
 
-// Key for offchain storage of aggregated derived public keys
-pub const AGGREGATED_PUBLIC_KEYS: &[u8] = b"dkg-metadata::public_key";
-
-// Key for offchain storage of aggregated derived public keys for genesis authorities
-pub const AGGREGATED_PUBLIC_KEYS_AT_GENESIS: &[u8] = b"dkg-metadata::genesis_public_keys";
-
-// Key for offchain storage of derived public key
-pub const SUBMIT_KEYS_AT: &[u8] = b"dkg-metadata::submit_keys_at";
-
-// Key for offchain storage of derived public key
-pub const SUBMIT_GENESIS_KEYS_AT: &[u8] = b"dkg-metadata::submit_genesis_keys_at";
-
-// Key for offchain storage of derived public key signature
-pub const OFFCHAIN_PUBLIC_KEY_SIG: &[u8] = b"dkg-metadata::public_key_sig";
-// Key for offchain signed proposals storage
-pub const OFFCHAIN_SIGNED_PROPOSALS: &[u8] = b"dkg-proposal-handler::signed_proposals";
+// Untrack interval for unsigned proposals completed stages for signing
 pub const UNTRACK_INTERVAL: u32 = 10;
 
 #[derive(Clone, Debug, PartialEq, Eq, codec::Encode, codec::Decode)]
@@ -84,33 +70,6 @@ pub struct AggregatedPublicKeys {
 impl<BlockNumber> Default for OffchainSignedProposals<BlockNumber> {
 	fn default() -> Self {
 		Self { proposals: Vec::default() }
-	}
-}
-
-pub mod offchain_crypto {
-	use sp_core::sr25519::Signature as Sr25519Signature;
-	use sp_runtime::{
-		app_crypto::{app_crypto, sr25519},
-		key_types::ACCOUNT,
-		traits::Verify,
-		MultiSignature, MultiSigner,
-	};
-	app_crypto!(sr25519, ACCOUNT);
-
-	pub struct OffchainAuthId;
-
-	impl frame_system::offchain::AppCrypto<MultiSigner, MultiSignature> for OffchainAuthId {
-		type RuntimeAppPublic = Public;
-		type GenericSignature = sp_core::sr25519::Signature;
-		type GenericPublic = sp_core::sr25519::Public;
-	}
-
-	impl frame_system::offchain::AppCrypto<<Sr25519Signature as Verify>::Signer, Sr25519Signature>
-		for OffchainAuthId
-	{
-		type RuntimeAppPublic = Public;
-		type GenericSignature = sp_core::sr25519::Signature;
-		type GenericPublic = sp_core::sr25519::Public;
 	}
 }
 
@@ -177,9 +136,13 @@ pub enum ConsensusLog<AuthorityId: Codec> {
 	/// MMR root hash.
 	#[codec(index = 3)]
 	MmrRoot(MmrRootHash),
-	/// The authority keys have changed
+	/// The DKG keys have changed
 	#[codec(index = 4)]
-	KeyRefresh { old_public_key: Vec<u8>, new_public_key: Vec<u8>, new_key_signature: Vec<u8> },
+	KeyRefresh {
+		old_public_key: Vec<u8>,
+		new_public_key: Vec<u8>,
+		new_key_signature: Vec<u8>
+	},
 }
 
 #[derive(Encode, Decode, PartialEq, Eq, Clone, RuntimeDebug, scale_info::TypeInfo)]
