+++ dkg-runtime-primitives/src/offchain/mod.rs
@@ -1,0 +1,2 @@
+pub mod storage_keys;
+pub mod crypto;
