+++ pallets/dkg-metadata/src/mock.rs
@@ -130,7 +130,7 @@ impl pallet_dkg_metadata::Config for Test {
 	type Event = Event;
 	type OnAuthoritySetChangeHandler = ();
 	type OnDKGPublicKeyChangeHandler = ();
-	type OffChainAuthId = dkg_runtime_primitives::offchain_crypto::OffchainAuthId;
+	type OffChainAuthId = dkg_runtime_primitives::offchain::crypto::OffchainAuthId;
 	type NextSessionRotation = pallet_session::PeriodicSessions<Period, Offset>;
 	type RefreshDelay = RefreshDelay;
 	type TimeToRestart = TimeToRestart;
