+++ pallets/dkg-proposal-handler/src/lib.rs
@@ -16,8 +16,8 @@ use dkg_runtime_primitives::{
 	Address, DKGPayloadKey, EIP1559TransactionMessage, EIP2930TransactionMessage,
 	LegacyTransactionMessage, OffchainSignedProposals, Proposal, ProposalAction,
 	ProposalHandlerTrait, ProposalHeader, ProposalKind, ProposalNonce, TransactionV2,
-	OFFCHAIN_SIGNED_PROPOSALS,
 };
+use dkg_runtime_primitives::offchain::storage_keys::OFFCHAIN_SIGNED_PROPOSALS;
 use frame_support::pallet_prelude::*;
 use frame_system::{
 	offchain::{AppCrypto, SendSignedTransaction, Signer},
